@@ -26,13 +26,13 @@
         }
     },
     "require": {
-        "php": ">=5.3.23",
-        "zendframework/zend-eventmanager": "~2.4.0",
-        "zendframework/zend-json": "~2.4.0",
-        "zendframework/zend-loader": "~2.4.0",
-        "zendframework/zend-mvc": "~2.4.0",
-        "zendframework/zend-paginator": "~2.4.0",
-        "zendframework/zend-stdlib": "~2.4.0",
+        "php": ">=5.5",
+        "zendframework/zend-eventmanager": "^2.4.0",
+        "zendframework/zend-json": "^2.4.0",
+        "zendframework/zend-loader": "^2.4.0",
+        "zendframework/zend-mvc": "^2.4.0",
+        "zendframework/zend-paginator": "^2.4.0",
+        "zendframework/zend-stdlib": "^2.4.0",
         "zfcampus/zf-api-problem": "~1.0",
         "zfcampus/zf-content-negotiation": "~1.0",
         "zfcampus/zf-hal": "~1.1",
@@ -40,15 +40,15 @@
     },
     "require-dev": {
         "phpunit/phpunit": "~4.7",
-        "squizlabs/php_codesniffer": "^2.3",
-        "zendframework/zend-console": "~2.4.0",
-        "zendframework/zend-escaper": "~2.4.0",
-        "zendframework/zend-http": "~2.4.0",
-        "zendframework/zend-inputfilter": "~2.4.0",
-        "zendframework/zend-servicemanager": "~2.4.0",
-        "zendframework/zend-uri": "~2.4.0",
-        "zendframework/zend-validator": "~2.4.0",
-        "zendframework/zend-view": "~2.4.0"
+        "squizlabs/php_codesniffer": "^2.3.1",
+        "zendframework/zend-console": "^2.4.0",
+        "zendframework/zend-escaper": "^2.4.0",
+        "zendframework/zend-http": "^2.4.0",
+        "zendframework/zend-inputfilter": "^2.4.0",
+        "zendframework/zend-servicemanager": "^2.4.0",
+        "zendframework/zend-uri": "^2.4.0",
+        "zendframework/zend-validator": "^2.4.0",
+        "zendframework/zend-view": "^2.4.0"
     },
     "autoload": {
         "psr-4": {
