@@ -18,6 +18,6 @@ class ArraySerializable
 
     public function getArrayCopy()
     {
-        return array('foo' => 'bar');
+        return ['foo' => 'bar'];
     }
 }
