@@ -15,6 +15,6 @@ class JsonSerializable implements JsonSerializableInterface
 {
     public function jsonSerialize()
     {
-        return array('foo' => 'bar');
+        return ['foo' => 'bar'];
     }
 }
